+++ packages/snaps-controllers/src/snaps/SnapController.ts
@@ -8,6 +8,7 @@ import type {
   ControllerStateChangeEvent,
 } from '@metamask/base-controller';
 import { BaseController } from '@metamask/base-controller';
+import type { CryptographicFunctions } from '@metamask/key-tree';
 import type {
   Caveat,
   GetEndowments,
@@ -476,7 +477,7 @@ export type SnapInstallFailed = {
  */
 export type SnapInstalled = {
   type: `${typeof controllerName}:snapInstalled`;
-  payload: [snap: TruncatedSnap, origin: string];
+  payload: [snap: TruncatedSnap, origin: string, preinstalled: boolean];
 };
 
 /**
@@ -500,7 +501,12 @@ export type SnapUnblocked = {
  */
 export type SnapUpdated = {
   type: `${typeof controllerName}:snapUpdated`;
-  payload: [snap: TruncatedSnap, oldVersion: string, origin: string];
+  payload: [
+    snap: TruncatedSnap,
+    oldVersion: string,
+    origin: string,
+    preinstalled: boolean,
+  ];
 };
 
 /**
@@ -706,7 +712,14 @@ type SnapControllerArgs = {
    * @returns The feature flags.
    */
   getFeatureFlags: () => DynamicFeatureFlags;
+
+  /**
+   * The cryptographic functions to use for the client. This may be an empty
+   * object to fall back to the default cryptographic functions.
+   */
+  clientCryptography?: CryptographicFunctions;
 };
+
 type AddSnapArgs = {
   id: SnapId;
   origin: string;
@@ -794,6 +807,8 @@ export class SnapController extends BaseController<
 
   #getFeatureFlags: () => DynamicFeatureFlags;
 
+  #clientCryptography: CryptographicFunctions | undefined;
+
   #detectSnapLocation: typeof detectSnapLocation;
 
   #snapsRuntimeData: Map<SnapId, SnapRuntimeData>;
@@ -827,6 +842,7 @@ export class SnapController extends BaseController<
     encryptor,
     getMnemonic,
     getFeatureFlags = () => ({}),
+    clientCryptography,
   }: SnapControllerArgs) {
     super({
       messenger,
@@ -882,6 +898,7 @@ export class SnapController extends BaseController<
     this.#encryptor = encryptor;
     this.#getMnemonic = getMnemonic;
     this.#getFeatureFlags = getFeatureFlags;
+    this.#clientCryptography = clientCryptography;
     this.#preinstalledSnaps = preinstalledSnaps;
     this._onUnhandledSnapError = this._onUnhandledSnapError.bind(this);
     this._onOutboundRequest = this._onOutboundRequest.bind(this);
@@ -1229,6 +1246,24 @@ export class SnapController extends BaseController<
       this.update((state) => {
         state.snaps[snapId].status = SnapStatus.Stopped;
       });
+
+      // Emit events
+      if (isUpdate) {
+        this.messagingSystem.publish(
+          'SnapController:snapUpdated',
+          this.getTruncatedExpect(snapId),
+          existingSnap.version,
+          'metamask',
+          true,
+        );
+      } else {
+        this.messagingSystem.publish(
+          'SnapController:snapInstalled',
+          this.getTruncatedExpect(snapId),
+          'metamask',
+          true,
+        );
+      }
     }
   }
 
@@ -1731,7 +1766,13 @@ export class SnapController extends BaseController<
 
     const salt = passedSalt ?? this.#encryptor.generateSalt();
     const mnemonicPhrase = await this.#getMnemonic();
-    const entropy = await getEncryptionEntropy({ snapId, mnemonicPhrase });
+
+    const entropy = await getEncryptionEntropy({
+      snapId,
+      mnemonicPhrase,
+      cryptographicFunctions: this.#clientCryptography,
+    });
+
     const encryptionKey = await this.#encryptor.keyFromPassword(
       entropy,
       salt,
@@ -2323,6 +2364,7 @@ export class SnapController extends BaseController<
           `SnapController:snapInstalled`,
           this.getTruncatedExpect(snapId),
           origin,
+          false,
         ),
       );
 
@@ -2332,6 +2374,7 @@ export class SnapController extends BaseController<
           this.getTruncatedExpect(snapId),
           oldVersion,
           origin,
+          false,
         ),
       );
 
@@ -2537,6 +2580,13 @@ export class SnapController extends BaseController<
   ): Promise<TruncatedSnap> {
     this.#assertCanInstallSnaps();
     this.#assertCanUsePlatform();
+
+    const snap = this.getExpect(snapId);
+
+    if (snap.preinstalled) {
+      throw new Error('Preinstalled Snaps cannot be manually updated.');
+    }
+
     if (!isValidSemVerRange(newVersionRange)) {
       throw new Error(
         `Received invalid snap version range: "${newVersionRange}".`,
@@ -2556,8 +2606,6 @@ export class SnapController extends BaseController<
         origin,
         true,
       );
-
-      const snap = this.getExpect(snapId);
 
       const oldManifest = snap.manifest;
 
@@ -2679,6 +2727,7 @@ export class SnapController extends BaseController<
           truncatedSnap,
           snap.version,
           origin,
+          false,
         );
       }