+++ packages/snaps-rpc-methods/src/restricted/getBip32Entropy.ts
@@ -1,3 +1,4 @@
+import type { CryptographicFunctions } from '@metamask/key-tree';
 import type {
   PermissionSpecificationBuilder,
   PermissionValidatorConstraint,
@@ -31,6 +32,15 @@ export type GetBip32EntropyMethodHooks = {
    * @returns A promise that resolves once the extension is unlocked.
    */
   getUnlockPromise: (shouldShowUnlockRequest: boolean) => Promise<void>;
+
+  /**
+   * Get the cryptographic functions to use for the client. This may return an
+   * empty object or `undefined` to fall back to the default cryptographic
+   * functions.
+   *
+   * @returns The cryptographic functions to use for the client.
+   */
+  getClientCryptography: () => CryptographicFunctions | undefined;
 };
 
 type GetBip32EntropySpecificationBuilderOptions = {
@@ -81,6 +91,7 @@ const specificationBuilder: PermissionSpecificationBuilder<
 const methodHooks: MethodHooksObject<GetBip32EntropyMethodHooks> = {
   getMnemonic: true,
   getUnlockPromise: true,
+  getClientCryptography: true,
 };
 
 export const getBip32EntropyBuilder = Object.freeze({
@@ -96,12 +107,15 @@ export const getBip32EntropyBuilder = Object.freeze({
  * @param hooks.getMnemonic - A function to retrieve the Secret Recovery Phrase of the user.
  * @param hooks.getUnlockPromise - A function that resolves once the MetaMask extension is unlocked
  * and prompts the user to unlock their MetaMask if it is locked.
+ * @param hooks.getClientCryptography - A function to retrieve the cryptographic
+ * functions to use for the client.
  * @returns The method implementation which returns a `JsonSLIP10Node`.
  * @throws If the params are invalid.
  */
 export function getBip32EntropyImplementation({
   getMnemonic,
   getUnlockPromise,
+  getClientCryptography,
 }: GetBip32EntropyMethodHooks) {
   return async function getBip32Entropy(
     args: RestrictedMethodOptions<GetBip32EntropyParams>,
@@ -115,6 +129,7 @@ export function getBip32EntropyImplementation({
       curve: params.curve,
       path: params.path,
       secretRecoveryPhrase: await getMnemonic(),
+      cryptographicFunctions: getClientCryptography(),
     });
 
     return node.toJSON();