+++ packages/snaps-sdk/src/jsx/validation.test.tsx
@@ -783,6 +783,19 @@ describe('CardStruct', () => {
       value="$1200"
       extra="0.12 ETH"
     />,
+    <Card
+      image="<svg />"
+      title={
+        <Address
+          address="0x1234567890123456789012345678901234567890"
+          displayName
+          avatar={false}
+        />
+      }
+      description="Description"
+      value="$1200"
+      extra="0.12 ETH"
+    />,
   ])('validates a card element', (value) => {
     expect(is(value, CardStruct)).toBe(true);
   });
@@ -1246,6 +1259,7 @@ describe('TextStruct', () => {
     <Text>
       Hello, <Bold>world</Bold>
     </Text>,
+    <Text size="sm">foo</Text>,
   ])('validates a text element', (value) => {
     expect(is(value, TextStruct)).toBe(true);
   });